Add new app.js for lambda and comments solved
