+++ backend/src/middlewares/errorHandler.js
@@ -1,5 +1,6 @@
-import logger from '../utils/logger';
 import HttpStatus from 'http-status-codes';
+
+import logger from '../utils/logger';
 import buildError from '../utils/buildError';
 
 /**
